+++ vmware_horizon.py
@@ -193,12 +193,13 @@ class Inventory:
         headers = self.access_token
         headers["Content-Type"] = 'application/json'
         json_data = json.dumps(farm_data)
-        response = requests.post(f'{self.url}rest/inventory/v1/farms', verify=False,  headers=headers, data=json_data)
+        response = requests.post(f'{self.url}/rest/inventory/v1/farms', verify=False,  headers=headers, data=json_data)
         if response.status_code == 400:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code == 401:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         elif response.status_code == 403:
@@ -228,7 +229,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code == 401:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         elif response.status_code == 403:
@@ -591,7 +593,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -615,7 +618,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -639,7 +643,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -663,7 +668,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -687,7 +693,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -715,7 +722,8 @@ class Inventory:
             error_message = (response.json())["error_messages"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -774,7 +782,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code == 403:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         elif response.status_code != 200:
@@ -854,7 +863,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 204:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -964,7 +974,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 204:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -988,7 +999,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 204:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -1012,7 +1024,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -1036,7 +1049,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -1060,7 +1074,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -1084,7 +1099,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -1108,7 +1124,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -1193,7 +1210,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 201:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -1238,7 +1256,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -1285,7 +1304,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -1393,7 +1413,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -1440,7 +1461,8 @@ class Inventory:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2379,7 +2401,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 204:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2400,7 +2423,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 204:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2663,7 +2687,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2684,7 +2709,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2704,7 +2730,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2724,7 +2751,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2744,7 +2772,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2764,7 +2793,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2784,7 +2814,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2795,7 +2826,7 @@ class Config:
             else:
                 return response.json()
 
-    def get_im_versions(self, im_version_id : str) -> dict:
+    def get_im_version(self, im_version_id : str) -> dict:
         """Gets image management version.
 
         Available for Horizon 7.12 and later."""
@@ -2804,7 +2835,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -2943,7 +2975,7 @@ class Config:
             else:
                 return response.status_code
 
-    def delete_im_asset(self, im_asset_id : str) -> dict:
+    def delete_im_asset(self, im_asset_id : str):
         """Deletes image management asset.
 
         Available for Horizon 7.12 and later."""
@@ -2952,7 +2984,8 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 204:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3042,7 +3075,7 @@ class Config:
             else:
                 return response.status_code
 
-    def delete_im_tag(self, im_tag_id : str) -> dict:
+    def delete_im_tag(self, im_tag_id : str):
         """Deletes image management tag.
 
         Available for Horizon 7.12 and later."""
@@ -3051,7 +3084,267 @@ class Config:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
+        elif response.status_code != 204:
             raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def new_im_stream(self,description : str,name : str, operating_system:str,publisher : str,source : str,status : str, 
+        additional_details_1:str = "", additional_details_2:str = "", additional_details_3:str = ""):
+        """Creates image management stream.
+
+        Requires all as string:
+            description
+            name
+            operating_system :  UNKNOWN, WINDOWS_XP, WINDOWS_VISTA, WINDOWS_7, WINDOWS_8, WINDOWS_10, WINDOWS_SERVER_2003,
+                                WINDOWS_SERVER_2008, WINDOWS_SERVER_2008_R2, WINDOWS_SERVER_2012, WINDOWS_SERVER_2012_R2,
+                                WINDOWS_SERVER_2016_OR_ABOVE, LINUX_OTHER, LINUX_SERVER_OTHER, LINUX_UBUNTU, LINUX_RHEL, LINUX_SUSE, LINUX_CENTOS
+            status : AVAILABLE, DELETED, DISABLED, FAILED, IN_PROGRESS, PARTIALLY_AVAILABLE, PENDING
+            publisher
+            source : MARKET_PLACE, UPLOADED, COPIED_FROM_STREAM, COPIED_FROM_VERSION
+            Optional:
+                additional_details_1
+                additional_details_2
+                additional_details_3
+        Available for Horizon 7.12 and later."""
+        valid_operating_system =    [ "UNKNOWN", "WINDOWS_XP", "WINDOWS_VISTA", "WINDOWS_7", "WINDOWS_8", "WINDOWS_10", "WINDOWS_SERVER_2003", 
+                                    "WINDOWS_SERVER_2008", "WINDOWS_SERVER_2008_R2", "WINDOWS_SERVER_2012", "WINDOWS_SERVER_2012_R2", "WINDOWS_SERVER_2016_OR_ABOVE", 
+                                    "LINUX_OTHER", "LINUX_SERVER_OTHER", "LINUX_UBUNTU", "LINUX_RHEL", "LINUX_SUSE", "LINUX_CENTOS" ]
+        valid_status = [ "AVAILABLE", "DELETED", "DISABLED", "FAILED", "IN_PROGRESS", "PARTIALLY_AVAILABLE", "PENDING" ]
+        valid_source = [ "MARKET_PLACE", "UPLOADED", "COPIED_FROM_STREAM", "COPIED_FROM_VERSION" ]
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        data = {}
+        additional_details = {}
+        additional_details["additionalProp1"] = additional_details_1
+        additional_details["additionalProp2"] = additional_details_2
+        additional_details["additionalProp3"] = additional_details_3
+        data["description"] = description
+        data["name"] = name
+        if operating_system in valid_operating_system:
+            data["operating_system"] = operating_system
+        else:
+            raise Exception(f"Error: please provide a valid operating_system from these options: {valid_operating_system}")
+        data["publisher"] = publisher
+        if source in valid_source:
+            data["source"] = source
+        else:
+            raise Exception(f"Error: please provide a valid source from these options: {valid_source}")
+        if status in valid_status:
+            data["status"] = status
+        else:
+            raise Exception(f"Error: please provide a valid status from these options: {valid_status}")
+        json_data = json.dumps(data)
+        response = requests.post(f'{self.url}/rest/config/v1/im-streams', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 409:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code != 201:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.status_code
+
+    def update_im_stream(self,im_stream_id : str, description : str,name : str, operating_system:str,publisher : str,source : str,status : str, 
+        additional_details_1:str = "", additional_details_2:str = "", additional_details_3:str = ""):
+        """Updates image management stream.
+
+        Requires all as string:
+            description
+            name
+            operating_system :  UNKNOWN, WINDOWS_XP, WINDOWS_VISTA, WINDOWS_7, WINDOWS_8, WINDOWS_10, WINDOWS_SERVER_2003,
+                                WINDOWS_SERVER_2008, WINDOWS_SERVER_2008_R2, WINDOWS_SERVER_2012, WINDOWS_SERVER_2012_R2,
+                                WINDOWS_SERVER_2016_OR_ABOVE, LINUX_OTHER, LINUX_SERVER_OTHER, LINUX_UBUNTU, LINUX_RHEL, LINUX_SUSE, LINUX_CENTOS
+            status : AVAILABLE, DELETED, DISABLED, FAILED, IN_PROGRESS, PARTIALLY_AVAILABLE, PENDING
+            publisher
+            source : MARKET_PLACE, UPLOADED, COPIED_FROM_STREAM, COPIED_FROM_VERSION
+            Optional:
+                additional_details_1
+                additional_details_2
+                additional_details_3
+        Available for Horizon 7.12 and later."""
+        valid_operating_system =    [ "UNKNOWN", "WINDOWS_XP", "WINDOWS_VISTA", "WINDOWS_7", "WINDOWS_8", "WINDOWS_10", "WINDOWS_SERVER_2003", 
+                                    "WINDOWS_SERVER_2008", "WINDOWS_SERVER_2008_R2", "WINDOWS_SERVER_2012", "WINDOWS_SERVER_2012_R2", "WINDOWS_SERVER_2016_OR_ABOVE", 
+                                    "LINUX_OTHER", "LINUX_SERVER_OTHER", "LINUX_UBUNTU", "LINUX_RHEL", "LINUX_SUSE", "LINUX_CENTOS" ]
+        valid_status = [ "AVAILABLE", "DELETED", "DISABLED", "FAILED", "IN_PROGRESS", "PARTIALLY_AVAILABLE", "PENDING" ]
+        valid_source = [ "MARKET_PLACE", "UPLOADED", "COPIED_FROM_STREAM", "COPIED_FROM_VERSION" ]
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        data = {}
+        additional_details = {}
+        additional_details["additionalProp1"] = additional_details_1
+        additional_details["additionalProp2"] = additional_details_2
+        additional_details["additionalProp3"] = additional_details_3
+        data["description"] = description
+        data["name"] = name
+        if operating_system in valid_operating_system:
+            data["operating_system"] = operating_system
+        else:
+            raise Exception(f"Error: please provide a valid operating_system from these options: {valid_operating_system}")
+        data["publisher"] = publisher
+        if source in valid_source:
+            data["source"] = source
+        else:
+            raise Exception(f"Error: please provide a valid source from these options: {valid_source}")
+        if status in valid_status:
+            data["status"] = status
+        else:
+            raise Exception(f"Error: please provide a valid status from these options: {valid_status}")
+        json_data = json.dumps(data)
+        response = requests.put(f'{self.url}/rest/config/v1/im-streams/{im_stream_id}', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 409:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code != 204:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.status_code
+
+    def delete_im_stream(self, im_stream_id : str):
+        """Deletes image management stream.
+
+        Available for Horizon 7.12 and later."""
+        response = requests.delete(f'{self.url}/rest/config/v1/im-streams/{im_stream_id}', verify=False,  headers=self.access_token)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 404:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
+        elif response.status_code != 204:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def new_im_version(self,description : str,name : str, im_stream_id : str,status : str, additional_details_1:str = "", additional_details_2:str = "", additional_details_3:str = ""):
+        """Creates image management version.
+
+        Requires all as string:
+            description
+            name
+            im_stream_id
+            status : AVAILABLE, DEPLOYING_VM, DEPLOYMENT_DONE, DELETED, DISABLED, FAILED, PARTIALLY_AVAILABLE, PUBLISHING, REPLICATING
+            Optional:
+                additional_details_1
+                additional_details_2
+                additional_details_3
+        Available for Horizon 7.12 and later."""
+        valid_status = [ "AVAILABLE", "DEPLOYING_VM", "DEPLOYMENT_DONE", "DELETED", "DISABLED", "FAILED", "PARTIALLY_AVAILABLE", "PUBLISHING", "REPLICATING" ]
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        data = {}
+        additional_details = {}
+        additional_details["additionalProp1"] = additional_details_1
+        additional_details["additionalProp2"] = additional_details_2
+        additional_details["additionalProp3"] = additional_details_3
+        data["description"] = description
+        data["im_stream_id"] = im_stream_id
+        data["name"] = name
+        if status in valid_status:
+            data["status"] = status
+        else:
+            raise Exception(f"Error: please provide a valid status from these options: {valid_status}")
+        json_data = json.dumps(data)
+        response = requests.post(f'{self.url}/rest/config/v1/im-versions', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 409:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code != 201:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.status_code
+
+    def update_im_version(self,description : str,name : str, im_version_id : str,status : str, additional_details_1:str = "", additional_details_2:str = "", additional_details_3:str = ""):
+        """Updates] image management version.
+
+        Requires all as string:
+            description
+            name
+            im_stream_id
+            status : AVAILABLE, DEPLOYING_VM, DEPLOYMENT_DONE, DELETED, DISABLED, FAILED, PARTIALLY_AVAILABLE, PUBLISHING, REPLICATING
+            Optional:
+                additional_details_1
+                additional_details_2
+                additional_details_3
+        Available for Horizon 7.12 and later."""
+        valid_status = [ "AVAILABLE", "DEPLOYING_VM", "DEPLOYMENT_DONE", "DELETED", "DISABLED", "FAILED", "PARTIALLY_AVAILABLE", "PUBLISHING", "REPLICATING" ]
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        data = {}
+        additional_details = {}
+        additional_details["additionalProp1"] = additional_details_1
+        additional_details["additionalProp2"] = additional_details_2
+        additional_details["additionalProp3"] = additional_details_3
+        data["description"] = description
+        data["name"] = name
+        if status in valid_status:
+            data["status"] = status
+        else:
+            raise Exception(f"Error: please provide a valid status from these options: {valid_status}")
+        json_data = json.dumps(data)
+        response = requests.put(f'{self.url}/rest/config/v1/im-versions/{im_version_id}', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 409:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code != 204:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.status_code
+
+    def delete_im_version(self, im_version_id : str):
+        """Deletes image management version.
+
+        Available for Horizon 7.12 and later."""
+        response = requests.delete(f'{self.url}/rest/config/v1/im-versions/{im_version_id}', verify=False,  headers=self.access_token)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 404:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 204:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3104,7 +3397,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3127,7 +3421,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3150,7 +3445,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3172,7 +3468,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3193,7 +3490,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3214,7 +3512,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3276,7 +3575,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3297,7 +3597,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3318,7 +3619,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3339,7 +3641,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3366,7 +3669,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3388,7 +3692,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3410,7 +3715,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3438,7 +3744,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3460,7 +3767,8 @@ class External:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3585,7 +3893,8 @@ class Entitlements:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3609,7 +3918,8 @@ class Entitlements:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3633,7 +3943,8 @@ class Entitlements:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3689,7 +4000,8 @@ class Entitlements:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3713,7 +4025,8 @@ class Entitlements:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3737,7 +4050,8 @@ class Entitlements:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3822,7 +4136,8 @@ class Entitlements:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3844,7 +4159,8 @@ class Entitlements:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3927,7 +4243,8 @@ class Entitlements:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else:
@@ -3949,7 +4266,8 @@ class Entitlements:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         elif response.status_code == 404:
-            raise Exception(f"Error {response.status_code}: {response.reason}")
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {response}")
         elif response.status_code != 200:
             raise Exception(f"Error {response.status_code}: {response.reason}")
         else: